requirements.txt
ultralytics
streamlit
pillow
python-dotenv
sambanova